move error_message
otherwise it will cause a circular import problem
